Implements basic order placement
Adds the basic structure for the user client and platform applications to enable order placement.
This includes setting up the event model, MQTT configuration, and API endpoints for creating orders.

The user client provides a REST endpoint for order creation which publishes to an MQTT topic.

Event model is extracted to own module, to allow sharing the event definition between user client and platform

Fixes #1